Crate monstorr

This is the main run-time crate for Monstorr. For more information about the tool, see its usage, which is also generated from doc comments inside this module. Starting at Arguments.

For more detailed documentation, including information on how to build a creature file, see [monstorr_lib].

Structs

Arguments

A tool for building D&D 5e creature stat blocks and other neat tricks

InputOutputData

A central structure for input data, since the same by so many commands. Note that output format is not specified, as that’s part of the command.

Enums

Command

Represents a command to be called by the main function

InputFormat

Represents in input format argument for commands that require input files

ListInputFormat

Represents a list input format argument for commands that require only list input files

OutputFormat

Represents a file output format argument for commands which produce statblocks

TemplateClass

Represents an argument requiring a class of templates for the ListTemplates command.

Functions

main

Parses the command line arguments, runs the tool, and prints out any error messages.

run

Runs the tool with the given arguments.