Enum monstorr::Command

pub(crate) enum Command {
    JSON {
        ugly: bool,
        input_output: InputOutputData,
    },
    MiniJinja {
        template: String,
        include: Vec<String>,
        input_output: InputOutputData,
    },
    HTML {
        input_output: InputOutputData,
        two_column: Option<usize>,
        fragment: bool,
    },
    Validate {
        input_output: InputOutputData,
    },
    ListTemplates {
        template_class: TemplateClass,
    },
    GetTemplate {
        name: String,
    },
    ListCreatures {
        format: ListInputFormat,
        input: Option<String>,
        type_: Option<String>,
        subtype: Option<String>,
        size: Option<String>,
        alignment: Option<String>,
        max_cr: Option<String>,
        min_cr: Option<String>,
    },
    MonstorrVersion,
}

Represents a command to be called by the main function

Variants

JSON

Generate a stat block in JSON format.

The structure for the stat-block JSON is documented in this tool’s code documentation. I hope to have a better link to this later.

Fields of JSON

ugly: bool

turns off pretty-printing of JSON text.

input_output: InputOutputData

MiniJinja

Run a stat block through a mini-jinja template to create almost any format you want.

Mini-Jinja is a template format related to another template format called Jinja. For more information on the syntax, see MiniJinja.

Currently, the code used to render these templates does not support auto-discovery of included template files. If your template needs these, you will need to include these using the ‘include’ option, below. Every attempt has been made to make sure that you can reference these files using relative paths in the include statement and still get your include to work correctly.

Fields of MiniJinja

template: String

a template file

include: Vec<String>

additional template file required by the main template (may be specified multiple times)

input_output: InputOutputData

HTML

Generate a stat block in html using some nice styles that resembles the official books.

This command utilizes built-in MiniJinja templates (see the mini-jinja command) to generate the HTML. It supports one or two-column formats, although the two-column format requires a height for the output box in px units. It can also generate just the div tag and its contents instead of the full HTML document.

If you wish to modify the output, retrieve the styles for embedding multiple div fragments in a page, or just reference them for how to write a template, use the list-templates html command to retrieve them. One simple template, for specifying the two-column mode, is generated at run-time, but a comment in the template explains how to add this yourself.

The HTML templates were based on styles used in statblock5e. That code was converted from “web components” into plain HTML, so it can support older browsers and not require JavaScript.

Fields of HTML

input_output: InputOutputDatatwo_column: Option<usize>

a two-column stat-block is produced, with the specified height in pixels. The stat-block is one column if not specified.

fragment: bool

only the stat-block div is produced, you would then import it into your own HTML page (see list-templates command to get the default styles)

Validate

Produce creature files unprocessed.

This can be used for two purposes: to retrieve the text of creatures from the ‘stored’ input, and to validate creature files without building. The latter has few practical uses, but the former is useful if you wish to see examples of real creature documents.

Attempting to use this on non-creature files will cause an error.

Fields of Validate

input_output: InputOutputData

ListTemplates

List built-in template files by template class, so you can modify or reference them.

Fields of ListTemplates

template_class: TemplateClass

The class of templates you want to list

GetTemplate

Retrieve the source for a template so you can save and customize it

Fields of GetTemplate

name: String

The name of the template you want to extract

ListCreatures

List monsters in a file, filtering for specific data.

This is a simple filtering tool to help you find creatures in a list. It currently only supports open5e-list files. See the information on input files in the main help for more information on this format.

Fields of ListCreatures

format: ListInputFormat

format of the input file (stored creatures can not be listed yet)

input: Option<String>

input file, if not specified will read from stdin

type_: Option<String>

type (humanoid, undead, monstrosity, etc.) of creatures to show

subtype: Option<String>

subtype (goblinoid, shapechanger, etc.) of creatures to show

size: Option<String>

size of creatures to show

alignment: Option<String>

alignment of creatures to show

max_cr: Option<String>

max challenge rating to show

min_cr: Option<String>

minimum challenge rating to show

MonstorrVersion

Display the creature format version used by this tool.

This version is used in the Monstorr command in creature files to make sure you’re processing your creature in the right version of Monstorr.